test(handler-graphql): disable introspection
